splitting gui setup into separate scripts
