+++ src/AV_Spex/utils/app_runner.py
@@ -1,0 +1,18 @@
+#!/usr/bin/env python3
+"""
+Main entry point for the AVSpex application.
+"""
+
+import sys
+from PyQt6.QtWidgets import QApplication
+from ui.main_window import MainWindow
+
+def main():
+    """Main application entry point."""
+    app = QApplication(sys.argv)
+    window = MainWindow()
+    window.show()
+    sys.exit(app.exec())
+
+if __name__ == "__main__":
+    main()
